+++ R/io.R
@@ -5,16 +5,18 @@
 #' non-UTF8 lines are found), and \code{write_utf8()} calls
 #' \code{writeLines(enc2utf8(text), useBytes = TRUE)}.
 #' @param con A connection or a file path.
+#' @param error Whether to signal an error when non-UTF8 characters are detected
+#'   (if \code{FALSE}, only a warning message is issued).
 #' @param text A character vector (will be converted to UTF-8 via
 #'   \code{\link{enc2utf8}()}).
 #' @param ... Other arguments passed to \code{\link{writeLines}()} (except
 #'   \code{useBytes}, which is \code{TRUE} in \code{write_utf8()}).
 #' @export
-read_utf8 = function(con) {
+read_utf8 = function(con, error = FALSE) {
   x = readLines(con, encoding = 'UTF-8', warn = FALSE)
   i = invalid_utf8(x)
   n = length(i)
-  if (n > 0) warning(
+  if (n > 0) (if (error) stop else warning)(
     if (is.character(con)) c('The file ', con, ' is not encoded in UTF-8. '),
     'These lines contain invalid UTF-8 characters: ',
     paste(c(head(i), if (n > 6) '...'), collapse = ', ')